requirements.txt
telethon>=1.28.5
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2

